requirements.txt
faiss-gpu
datasets
numpy
pytorch-lightning
pyyaml
sentence-transformers
torch
tqdm
transformers
wandb
